chatbot_db/pyproject.toml
[project]
name = "chatbot-db"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "chromadb[server]==0.5.0",
    "openai==1.44.1",
    "python-dotenv==1.0.1",
]
